Renames five module to compat (#745)
* Renames five module to compat and updates all refs

# Conflicts:
#	CONTRIBUTING.rst
#	rasterio/plot.py
