+++ rasterio/compat.py
