+++ rasterio/crs.py
@@ -14,7 +14,7 @@ import json
 
 from rasterio._base import is_geographic_crs, is_projected_crs, is_same_crs
 from rasterio.errors import CRSError
-from rasterio.five import string_types
+from rasterio.compat import string_types
 
 
 def is_valid_crs(crs):
