+++ rasterio/env.py
@@ -6,7 +6,7 @@ from rasterio._drivers import (
     GDALEnv, del_gdal_config, get_gdal_config, set_gdal_config)
 from rasterio.dtypes import check_dtype
 from rasterio.errors import EnvError
-from rasterio.five import string_types
+from rasterio.compat import string_types
 from rasterio.transform import guard_transform
 from rasterio.vfs import parse_path, vsi_path
